refactor - improved names and typing for database query, added docstring, try/catch
